Make IdsApiObjectAsset non-abstract
The class does not feature any abstract methods so doesn't make sense for it to be abstract. Class can also be instantiated during the execution of IdsApiObject::factory which is not possible for an abstract class (and will cause an error if that code condition is reached).
c.f. 51b46567e6e9723b0a61c450c180f6768c8b56b9 from the drupal module.